bazel: further progress on the rule
* adding rule to extract class files from built jar
* adding gazelle java generation
* mutate passes but logs are not collected
