+++ .bazelrc
@@ -10,6 +10,14 @@ common --enable_bzlmod
 # https://bazelbuild.slack.com/archives/C014RARENH0/p1691158021917459?thread_ts=1691156601.420349&cid=C014RARENH0
 common --check_direct_dependencies=off
 
+
+common --java_runtime_version=remotejdk_17
+common --java_language_version=17
+common --tool_java_runtime_version=remotejdk_17
+common --tool_java_language_version=17
+test --test_output=errors
+common  --incompatible_strict_action_env
+
 # Load any settings specific to the current user.
 # .bazelrc.user should appear in .gitignore so that settings are not shared with team members
 # This needs to be last statement in this
