+++ MODULE.bazel
@@ -14,6 +14,8 @@ bazel_dep(name = "bazel_skylib_gazelle_plugin", version = "1.4.1", dev_dependenc
 bazel_dep(name = "aspect_bazel_lib", version = "1.32.1", dev_dependency = True)
 bazel_dep(name = "buildifier_prebuilt", version = "6.1.2", dev_dependency = True)
 
+bazel_dep(name = "contrib_rules_jvm", version = "0.24.0")
+
 bazel_dep(name = "rules_java", version = "7.4.0")
 bazel_dep(name = "rules_jvm_external", version = "6.0")
 
@@ -26,7 +28,7 @@ maven.install(
         "org.pitest:pitest-command-line:1.15.8",
         "org.pitest:pitest-junit5-plugin:1.2.1",
     ],
-    fetch_sources = False,
+    fetch_sources = True,
     lock_file = "//:maven_install.json",
 )
 use_repo(
