--- e2e/smoke/sample-from-upstream/src/main/java/com/example/BUILD.bazel
@@ -1,23 +1,0 @@
-load("@rules_java//java:defs.bzl", "java_library")
-load("@com_bookingcom_rules_pitest//pitest:defs.bzl", "java_pitest_test")
-
-java_library(
-    name = "example",
-    srcs = [
-        "CustomRecord.java",
-        "NoDataRecord.java",
-        "NotARecord.java",
-        "PureRecord.java",
-        "RecordWithCustomConstructor.java",
-        "RecordWithCustomEquals.java",
-        "RecordWithCustomHashCode.java",
-        "RecordWithCustomToString.java",
-    ],
-    visibility = ["//:__subpackages__"],
-)
-
-filegroup(
-    name = "java_srcs",
-    srcs = glob(["*.java"]),
-    visibility = ["//:__subpackages__"],
-)
