+++ e2e/smoke/sample-from-upstream/src/test/java/com/example/test/JUnitParamsTest.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.test;
 
 import static junitparams.JUnitParamsRunner.$;
