+++ e2e/smoke/sample-from-upstream/src/test/java/com/example/test/TheoryTest.java
@@ -12,7 +12,7 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and limitations under the License.
  */
-package com.example;
+package com.example.test;
 
 import org.junit.experimental.theories.DataPoint;
 import org.junit.experimental.theories.Theories;
