+++ e2e/smoke/test-helpers/src/test/java/com/example/test/CoveredByABeforeAfterClass.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.test;
 
 public class CoveredByABeforeAfterClass {
   public static int returnOne() {
