+++ e2e/smoke/test-helpers/src/test/java/com/example/test/CoveredByJUnitThreeSuite.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.test;
 
 public class CoveredByJUnitThreeSuite {
