+++ e2e/smoke/test-helpers/src/test/java/com/example/test/LineNumbersSpanBlocks.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.test;
 
 //@formatter:off
 public class LineNumbersSpanBlocks {
