+++ maven_install.json
@@ -1,71 +1,82 @@
 {
   "__AUTOGENERATED_FILE_DO_NOT_MODIFY_THIS_FILE_MANUALLY": "THERE_IS_NO_DATA_ONLY_ZUUL",
   "__INPUT_ARTIFACTS_HASH": 317352002,
-  "__RESOLVED_ARTIFACTS_HASH": -1899371614,
+  "__RESOLVED_ARTIFACTS_HASH": 1540523952,
   "artifacts": {
     "org.apache.commons:commons-lang3": {
       "shasums": {
-        "jar": "d919d904486c037f8d193412da0c92e22a9fa24230b9d67a57855c5c31c7e94e"
+        "jar": "d919d904486c037f8d193412da0c92e22a9fa24230b9d67a57855c5c31c7e94e",
+        "sources": "325a4551eee7d99f7616aa05b00ee3ca9d0cdc8face1b252a9864f2d945c58b3"
       },
       "version": "3.12.0"
     },
     "org.apache.commons:commons-text": {
       "shasums": {
-        "jar": "770cd903fa7b604d1f7ef7ba17f84108667294b2b478be8ed1af3bffb4ae0018"
+        "jar": "770cd903fa7b604d1f7ef7ba17f84108667294b2b478be8ed1af3bffb4ae0018",
+        "sources": "e767596163da9b22e8fc83ff954e41e19e96744df7625b06f8daac3c4d60f79d"
       },
       "version": "1.10.0"
     },
     "org.apiguardian:apiguardian-api": {
       "shasums": {
-        "jar": "b509448ac506d607319f182537f0b35d71007582ec741832a1f111e5b5b70b38"
+        "jar": "b509448ac506d607319f182537f0b35d71007582ec741832a1f111e5b5b70b38",
+        "sources": "277a7a4315412817beb6655b324dc7276621e95ebff00b8bf65e17a27b685e2d"
       },
       "version": "1.1.2"
     },
     "org.junit.platform:junit-platform-commons": {
       "shasums": {
-        "jar": "624a3d745ef1d28e955a6a67af8edba0fdfc5c9bad680a73f67a70bb950a683d"
+        "jar": "624a3d745ef1d28e955a6a67af8edba0fdfc5c9bad680a73f67a70bb950a683d",
+        "sources": "35ad4740fd4388406044d288dac4059f01c484f30e2b2a931703aae9dbca08c3"
       },
       "version": "1.9.2"
     },
     "org.junit.platform:junit-platform-engine": {
       "shasums": {
-        "jar": "25f23dc535a091e9dc80c008faf29dcb92be902e6911f77a736fbaf019908367"
+        "jar": "25f23dc535a091e9dc80c008faf29dcb92be902e6911f77a736fbaf019908367",
+        "sources": "31906715fb92e10b96da6cdcd1c75066bd275e961949fcd0ab61b7501c867eb9"
       },
       "version": "1.9.2"
     },
     "org.junit.platform:junit-platform-launcher": {
       "shasums": {
-        "jar": "eef139eb09c98e9cdd358b6ce4c6cdd59b30c6a88096e369c33ba96e67edf0e4"
+        "jar": "eef139eb09c98e9cdd358b6ce4c6cdd59b30c6a88096e369c33ba96e67edf0e4",
+        "sources": "3cf6e39c88b59e63851ac1981ce567a24ab89f89f7eb738145ad72dcba105b4a"
       },
       "version": "1.9.2"
     },
     "org.opentest4j:opentest4j": {
       "shasums": {
-        "jar": "58812de60898d976fb81ef3b62da05c6604c18fd4a249f5044282479fc286af2"
+        "jar": "58812de60898d976fb81ef3b62da05c6604c18fd4a249f5044282479fc286af2",
+        "sources": "b63495ef700fb2af2cdee8dd68659b27822650058234a602f9ed1d14b909a1a8"
       },
       "version": "1.2.0"
     },
     "org.pitest:pitest": {
       "shasums": {
-        "jar": "c4f8024a2faf52f3cf5d205b47b1e18fe3cdc30503aaa9ecf75fe67cf3b0ff76"
+        "jar": "c4f8024a2faf52f3cf5d205b47b1e18fe3cdc30503aaa9ecf75fe67cf3b0ff76",
+        "sources": "c84829cf44718001760faa14f12730043418cc38d95dd6f304323fd353c54536"
       },
       "version": "1.15.8"
     },
     "org.pitest:pitest-command-line": {
       "shasums": {
-        "jar": "f1126c5b834b3c8c09f5d6c44a4e38929001283886f2328beaf6e12132d0af2e"
+        "jar": "f1126c5b834b3c8c09f5d6c44a4e38929001283886f2328beaf6e12132d0af2e",
+        "sources": "d35f679d2d8572eef5bdc36f786efa046076084647a3e8b057908f09e66f9910"
       },
       "version": "1.15.8"
     },
     "org.pitest:pitest-entry": {
       "shasums": {
-        "jar": "6cd71474c6bef7e05d7d53086253ded57c9f9a2e636ee99c39cfe96ec5131752"
+        "jar": "6cd71474c6bef7e05d7d53086253ded57c9f9a2e636ee99c39cfe96ec5131752",
+        "sources": "2b6029f31e8d477f5ce279f5b54c1cc901162fd8d0ad06e6b79e845dfb9f185f"
       },
       "version": "1.15.8"
     },
     "org.pitest:pitest-junit5-plugin": {
       "shasums": {
-        "jar": "9d944887a583ebd4adc33d2d26b8b56639fd53fee93b4d068a90596705c072b5"
+        "jar": "9d944887a583ebd4adc33d2d26b8b56639fd53fee93b4d068a90596705c072b5",
+        "sources": "16f0be3a02acc92b562d0a23b8f1cf36c3599f14e4107d4050b6fa41719619c9"
       },
       "version": "1.2.1"
     }
@@ -271,16 +282,27 @@
   "repositories": {
     "https://repo1.maven.org/maven2/": [
       "org.apache.commons:commons-lang3",
+      "org.apache.commons:commons-lang3:jar:sources",
       "org.apache.commons:commons-text",
+      "org.apache.commons:commons-text:jar:sources",
       "org.apiguardian:apiguardian-api",
+      "org.apiguardian:apiguardian-api:jar:sources",
       "org.junit.platform:junit-platform-commons",
+      "org.junit.platform:junit-platform-commons:jar:sources",
       "org.junit.platform:junit-platform-engine",
+      "org.junit.platform:junit-platform-engine:jar:sources",
       "org.junit.platform:junit-platform-launcher",
+      "org.junit.platform:junit-platform-launcher:jar:sources",
       "org.opentest4j:opentest4j",
+      "org.opentest4j:opentest4j:jar:sources",
       "org.pitest:pitest",
       "org.pitest:pitest-command-line",
+      "org.pitest:pitest-command-line:jar:sources",
       "org.pitest:pitest-entry",
-      "org.pitest:pitest-junit5-plugin"
+      "org.pitest:pitest-entry:jar:sources",
+      "org.pitest:pitest-junit5-plugin",
+      "org.pitest:pitest-junit5-plugin:jar:sources",
+      "org.pitest:pitest:jar:sources"
     ]
   },
   "version": "2"
